+++ src/client/java/com/launium/ghostify/client/ui/container/AutoPetNotificationContainer.java
@@ -43,7 +43,7 @@ public class AutoPetNotificationContainer implements IContainer {
         Font font = Minecraft.getInstance().font;
         Easy2D.drawScreenText(font, warningText,
                 (left + right - ((AccessFont) font).getSplitter().stringWidth(warningText)) * 0.5F,
-                (top + bottom - (float) font.lineHeight) * 0.5F,
+                (top + bottom - font.lineHeight) * 0.5F,
                 Easy2D.TEXT_DEFAULT_COLOR, false);
     }
 }
